Spring security project completed

+++ src/main/java/com/keval/SpringSecurity/controller/AssignmentController.java
@@ -1,10 +1,10 @@
 package com.keval.SpringSecurity.controller;
 
+import java.time.LocalDate;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
@@ -14,7 +14,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
-import com.keval.SpringSecurity.config.MySecurityConfig;
 import com.keval.SpringSecurity.model.Assignment;
 import com.keval.SpringSecurity.model.User;
 import com.keval.SpringSecurity.model.UserDetailImpl;
@@ -25,18 +24,25 @@ import com.keval.SpringSecurity.service.AssignmentService;
 public class AssignmentController {
 	@Autowired
 	AssignmentService assignmentServiceImpl; 
-	@Autowired
-	private MySecurityConfig mySecurityConfig;
 	
 	@RequestMapping("/view")
 	public String viewAssignment(Model model) {
 		model.addAttribute("assignmentList",assignmentServiceImpl.getAssignments());
-		return "../ViewAssignment.jsp";
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		if (authentication != null && authentication.getPrincipal() instanceof UserDetails) {
+		    UserDetailImpl userDetails = (UserDetailImpl) authentication.getPrincipal();
+		    User user = userDetails.getUser();
+		    if((user.getRole().equals("ROLE_TEACHER"))) {
+		    	return "../ViewAssignment.jsp";
+		    }
+		}
+		return "../StudentAssignment.jsp";
 	}
 	
 	@PreAuthorize("hasRole('TEACHER')")
 	@RequestMapping("/create")
 	public String createPage(Model model) {
+		model.addAttribute("todayDate",LocalDate.now().toString());
 		model.addAttribute("assignment",new Assignment());
 		return "../AssignmentEntry.jsp";
 	}
@@ -51,6 +57,12 @@ public class AssignmentController {
 	@PreAuthorize("hasRole('TEACHER')")
 	@RequestMapping("/delete{assignmentId}")
 	public String Assignment(@PathVariable int assignmentId) {
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		UserDetailImpl userDetails = (UserDetailImpl) authentication.getPrincipal();
+	    User user = userDetails.getUser();
+	    if(!(user.getId()==assignmentServiceImpl.getAssignment(assignmentId).getCreatedBy())) {
+	    	return "../error.jsp";
+	    }
 		assignmentServiceImpl.deleteAssignment(assignmentId);
 		return "redirect:/assignment/view";
 	}
@@ -59,14 +71,18 @@ public class AssignmentController {
 	@RequestMapping(value="/submit", method= RequestMethod.POST)
 	public String createAssignment(@ModelAttribute() Assignment assignment) {
 		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-		if (authentication != null && authentication.getPrincipal() instanceof UserDetails) {
-		    UserDetailImpl userDetails = (UserDetailImpl) authentication.getPrincipal();
-		    User user = userDetails.getUser();
-		    if(!(user.getId()==assignment.getCreatedBy())) {
+		UserDetailImpl userDetails = (UserDetailImpl) authentication.getPrincipal();
+	    User user = userDetails.getUser();
+	    
+		if(assignment.getId()==0) {
+			assignment.setCreatedBy(user.getId());
+			assignmentServiceImpl.createAssignment(assignment);
+		}else {
+			if(!(user.getId()==assignment.getCreatedBy())) {
 		    	return "../error.jsp";
-		    }
+			}
+			assignmentServiceImpl.updateAssignment(assignment);
 		}
-		assignmentServiceImpl.updateAssignment(assignment);
 		return "redirect:/assignment/view";
 	}
 }
